Kernel module code cleanup
